Move calculateBestLoanValue to global helpers
