+++ src/pages/LendPage/components/PlaceOfferTab/PlaceProOffer/hooks/usePlaceProOffer.tsx
@@ -3,10 +3,9 @@ import { useEffect, useMemo } from 'react'
 import { useWallet } from '@solana/wallet-adapter-react'
 import { isEmpty } from 'lodash'
 
-import { useSolanaBalance } from '@banx/utils'
+import { formatDecimal, useSolanaBalance } from '@banx/utils'
 
-import { calculateBestLoanValue } from '../../PlaceLiteOffer/helpers'
-import { calculateOfferSize, getOfferErrorMessage } from '../../helpers'
+import { calculateBestLoanValue, calculateOfferSize, getOfferErrorMessage } from '../../helpers'
 import { OfferParams } from '../../hooks'
 import { useOfferTransactions } from '../../hooks/useOfferTransactions'
 import { useOfferFormController } from './useOfferFormController'
@@ -68,7 +67,7 @@ export const usePlaceProOffer = ({
 
     if (hasSolanaBalance && isNotEditMode && connected && !isEmpty(marketPreview)) {
       const bestLoanValue = calculateBestLoanValue(solanaBalance, marketPreview.bestOffer)
-      onLoanValueChange(bestLoanValue)
+      onLoanValueChange(formatDecimal(bestLoanValue))
     }
   }, [marketPreview, connected, solanaBalance, syntheticOffer, onLoanValueChange])
