+++ src/pages/LendPage/components/PlaceOfferTab/helpers.ts
@@ -118,3 +118,13 @@ export const calculateOfferSize: CalculateOfferSize = ({
   const offerSize = updatedBondOffer.fundsSolOrTokenBalance
   return offerSize
 }
+
+const TRANSACTION_FEE_IN_SOL = 0.01 //? transaction fee for prevent any case with not enough sol
+export const calculateBestLoanValue = (solanaBalance: number, bestOffer: number) => {
+  const balanceAfterDeductingFee = solanaBalance - TRANSACTION_FEE_IN_SOL
+  const maxLoanValue = Math.max(balanceAfterDeductingFee, 0)
+
+  const bestOfferInSol = bestOffer / 1e9
+  const bestLoanValue = Math.min(maxLoanValue, bestOfferInSol) || 0
+  return bestLoanValue
+}
